refactor: 종속적인 Modal 컴포넌트 폴더 정리
