+++ src/pages/UpdateTimeDivider/components/DoneFormModal.jsx
@@ -1,4 +1,4 @@
-import FormModal from './FormModal'
+import FormModal from '../../../shared/components/FormModal'
 import React from 'react'
 import { useRecoilState } from 'recoil'
 import { doneMode, mergeMode, modeState, originIdState, timerState } from 'state/timer'
